% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComboMutPlot.R
\name{ComboMutPlot}
\alias{ComboMutPlot}
\title{Create The Mutation Plot}
\usage{
ComboMutPlot(ReturnClass)
}
\arguments{
\item{ReturnClass}{instance of MutationPlot class}
}
\description{
Create The Mutation Plot
}
\examples{
mutationPlot(Data, "SampleName", "GeneName", "Info")
}
\keyword{Mutation}
\keyword{Waterfall}
