% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationPlot.R
\docType{data}
\name{Data}
\alias{Data}
\title{Create The Mutation Plot}
\format{An object of class \code{data.frame} with 5 rows and 3 columns.}
\usage{
Data
}
\arguments{
\item{Data}{Data frame of mutation data}

\item{SampleCol}{string - column name for samples}

\item{GeneCol}{string - column name for gene names}

\item{InfoCol}{string - column name for variant info (used to colour bar charts)}

\item{FuncColour}{list - optional colours specified for the columns}
}
\description{
Create The Mutation Plot
}
\examples{
mutationPlot(Data, "SampleName", "GeneName", "Info")
}
\keyword{Mutation}
\keyword{Waterfall}
