% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPlotClass.R
\docType{class}
\name{MutationPlot-class}
\alias{MutationPlot-class}
\alias{MutationPlot}
\title{Run the whole pipeline start to finish
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will create an instance of the CNVvault class
Then it will order this class based on the total amount of the geneome editted.
Then it will generate the positions for each sample on the y-axis
Then it will plot the output}
\arguments{
\item{mutationData}{dataframe of variant calls}

\item{plots}{Description}

\item{geneList}{Description}

\item{plotData}{Description}
}
\description{
Run the whole pipeline start to finish
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will create an instance of the CNVvault class
Then it will order this class based on the total amount of the geneome editted.
Then it will generate the positions for each sample on the y-axis
Then it will plot the output
}
\examples{
Mut <- MutationPlot()

}
\keyword{Mutation}
\keyword{Waterfall}
