% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneCount-pegasus.R, R/plotGeneCount.R
\name{plotGeneCount}
\alias{plotGeneCount}
\title{plotGeneCount function}
\usage{
plotGeneCount(ReturnClass, Colour_List = FALSE)

plotGeneCount(ReturnClass, Colour_List = FALSE)
}
\arguments{
\item{ReturnClass}{instance of MutationPlot class}

\item{Colour_list}{list - optional colours specified for the columns}
}
\description{
plotGeneCount function

plotGeneCount function
}
\examples{
## Pass the function a MuationPlot object
plotGeneCount(MutationPlot)
## Pass the function a MuationPlot object
plotGeneCount(MutationPlot)
}
\keyword{Mutation}
\keyword{Waterfall}
