% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleCount.R
\name{plotSampleCount}
\alias{plotSampleCount}
\title{plotSampleCount function}
\usage{
plotSampleCount(ReturnClass, Colour_List = FALSE)
}
\arguments{
\item{ReturnClass}{Description}

\item{Colour_List}{list - optional colours specified for the columns}
}
\description{
plotSampleCount function
}
\examples{
## Pass the function a MuationPlot object
plotSampleCount(MutationPlot)
}
\keyword{Mutation}
\keyword{Waterfall}
