% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterfall.R
\name{plotWaterfall}
\alias{plotWaterfall}
\title{plotWaterfall function}
\usage{
plotWaterfall(ReturnClass)
}
\arguments{
\item{ReturnClass}{instance of MutationPlot class}
}
\description{
plotWaterfall function
}
\examples{
## Pass the function a MuationPlot object
plotWaterfall(MutationPlot)
}
\keyword{Mutation}
\keyword{Waterfall}
