% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMutData.R
\name{prepareMutData}
\alias{prepareMutData}
\title{prepareMutData function}
\usage{
prepareMutData(Data, SampleCol = "Sample", GeneCol = "Gene", InfoCol = "Info")
}
\arguments{
\item{Data}{Data frame of mutation data}

\item{SampleCol}{string - column name for samples}

\item{GeneCol}{string - column name for gene names}

\item{InfoCol}{string - column name for variant info (used to colour bar charts)}
}
\description{
prepareMutData function
}
\examples{
## Pass the function a dataframe of mutation calls, and the names of the columns to use.
## Need the following columns: Sample, Gene, InfoCol
plotWaterfall(MutationPlot)
plotWaterfall(MutationPlot, SampleCol="Sample", GeneCol="Gene", InfoCol="Info")
}
\keyword{Mutation}
\keyword{Waterfall}
