% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutPlotClass.R
\docType{class}
\name{MutationPlot-class}
\alias{MutationPlot-class}
\alias{MutationPlot}
\title{Run the whole pipeline start to finish}
\arguments{
\item{mutationData}{dataframe of variant calls}

\item{plots}{Description}

\item{geneList}{Description}

\item{plotData}{Description}
}
\description{
Run the whole pipeline start to finish
}
\examples{
Mut <- MutationPlot()

}
\keyword{Mutation}
\keyword{Waterfall}
